feat(storyblok-ui): move LocaleRedirectProxy plugin into the package
Auto-loaded from packages/storyblok-ui/plugins/ for any project
depending on @graphcommerce/storyblok-ui. ifConfig: 'storyblok'
guards activation. Removed from the example since it's now
inherited transitively.
